% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bmi.R
\name{get_bmi}
\alias{get_bmi}
\title{Calculate body mass index}
\usage{
get_bmi(wt, ht, wt_units = "kg", ht_units = "cm")
}
\arguments{
\item{wt}{bodyweight}

\item{ht}{height}

\item{wt_units}{units of measure for \code{wt}. Can be quoted (recommended)
or unquoted. Default is kg}

\item{ht_units}{units of measure for \code{ht}. Can be quoted (recommended)
or unquoted. Default is cm}
}
\value{
An S3 object with class \code{units}
}
\description{
Calculate body mass index
}
\examples{
get_bmi(75, 180)
}
