% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b_package_doc.R
\docType{package}
\name{PAutilities}
\alias{PAutilities}
\title{anthropometry: Conveniently Characterize Body Measures}
\description{
This package provides an easy to use interface for recording and operating on
anthropometric values such as height, weight, body mass index (including
percentiles), fat mass, and fat-free mass.
}
