% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_convert.R
\name{unit_convert}
\alias{unit_convert}
\title{Flexibly convert anthropometric values between units}
\usage{
unit_convert(x, from, to)
}
\arguments{
\item{x}{a value or vector to convert}

\item{from}{current units. Can be quoted (recommended) or unquoted}

\item{to}{desired units. Can be quoted (recommended) or unquoted}
}
\value{
an S3 object with class \code{units}
}
\description{
Flexibly convert anthropometric values between units
}
\examples{
unit_convert(1, "in", "cm")
}
\seealso{
\link[units]{set_units}
}
