% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_status__main.R
\name{weight_status}
\alias{weight_status}
\alias{weight_status_adult}
\alias{weight_status_cdc_youth}
\alias{weight_status_who_youth}
\alias{weight_status_custom}
\title{Classify weight status}
\usage{
## Wrapper function:

  weight_status(
    bmi, method = c("adult", "CDC youth", "WHO youth", "custom"),
    select = c("bmi", "percentile", "classification", "cutoff", "all"),
    wt = NULL, ht = NULL, wt_units = "kg", ht_units = "cm",
    ...
  )


## Internal applicator functions that the wrapper calls
## based on the value of the `method` argument:

  weight_status_adult(bmi)

  weight_status_cdc_youth(
    bmi, sex = c("Male", "Female"),
    age_yrs = NULL, age_mos = NULL,
    severe_35 = TRUE
  )

  weight_status_who_youth(
    bmi, sex = c("Male", "Female"),
    age_yrs = NULL, age_mos = NULL,
    severe_35 = TRUE
  )

  weight_status_custom(
    bmi, cutoffs = c(-Inf, 18.5, 25, 30, 35, 40, Inf),
    labels = c("Underweight", "Healthy Weight", "Overweight",
      "Class 1 Obese", "Class 2 Obese", "Class 3 Obese"),
    right = FALSE, check_cutoffs = TRUE
  )
}
\arguments{
\item{bmi}{body mass index value(s). Must return \code{TRUE} when tested with
\code{\link{is_bmi}}. See details}

\item{method}{character scalar. The desired classification method. Options
are \code{"adult"} (the default), \code{"CDC youth"}, \code{"WHO youth"},
and \code{"custom"}}

\item{select}{character. The desired output. Can be \code{"bmi"},
\code{"percentile"}, \code{"classification"}, \code{"cutoff"} (i.e., the
severe obesity cutoff for youth), \code{"all"}, or a combination thereof.
Percentiles and severe obesity cutoffs will not be returned when
\code{method} is \code{adult} or \code{custom}, and an error will be thrown
if no other selections are made for those methods}

\item{wt}{body mass value(s)}

\item{ht}{height value(s)}

\item{wt_units}{units of measure for \code{wt}. Can be quoted (recommended)
or unquoted. Default is kg}

\item{ht_units}{units of measure for \code{ht}. Can be quoted (recommended)
or unquoted. Default is cm}

\item{...}{arguments passed to the internal applicator that corresponds with
\code{method}, which may include some of the the arguments below, depending
on the method}

\item{sex}{character. Sex of participant(s) or patient(s)}

\item{age_yrs}{age in years}

\item{age_mos}{age in months (optional)}

\item{severe_35}{logical. Include \eqn{BMI \ge 35} as a criterion for severe
obesity (per
\href{https://www.ahajournals.org/doi/full/10.1161/CIR.0b013e3182a5cfb3}{Kelly
 et al., 2013, \emph{Circulation}}), or not (per
\href{https://academic.oup.com/ajcn/article/90/5/1314/4598130}{Flegal et
al., 2009, \emph{Am J Clin Nutr}})}

\item{cutoffs}{the BMI cutoffs to use when determining weight status for
\code{weight_status_custom}. Default is the same cutoffs used for
\code{weight_status_adult}}

\item{labels}{character labels to assign for intervals based on
\code{cutoffs}. Must be either \code{NULL} or else a character vector of
length one less than \code{length(cutoffs)}. See ?cut for more information}

\item{right}{See ?cut}

\item{check_cutoffs}{logical. Should custom cutoffs be checked, with warnings
issued when potential issues are detected?}
}
\value{
A tibble indicating the BMI(s), percentile(s), classification(s),
  and/or severe obesity cutoff(s)
}
\description{
Classify weight status
}
\details{
Two options are available for passing in values. The first is to
  supply a value for \code{bmi}, which must return \code{TRUE} when tested
  with \code{\link{is_bmi}}. If attempting to pass in BMI as a numeric value,
  it can be wrapped in \code{\link{as_bmi}} to ensure success. The second
  option is to pass weight and height in as separate values, which will be
  fed into \code{\link{get_bmi}} to determine the value for further
  classification. Regardless of how values are passed in, careful attention
  should be given to ensure units are correct. The code and internal checks
  have been set up to help with this as much as possible, but their
  unavoidable limitations leave a good amount of onus on the user.

When \code{method} is \code{"CDC youth"} or \code{"WHO youth"}, only one of
\code{age_mos} and \code{age_yrs} is required. The former will be used if
both are provided. If \code{age_mos} is \emph{not} provided, it will be
calculated based on \code{age_yrs}, assuming 365.2425 days per year and
30.4375 days per month. Depending on how the initial age calculation was
made, rounding error will occur. Thus, use of the \code{\link{get_age}}
function is recommended, with \code{units = "months"}.
}
\examples{

set.seed(610)

weight_status(as_bmi(c(18, 20, 26, 31, 36, 42)))

weight_status(
  method = "CDC youth",
  bmi = as_bmi(rnorm(50, 23, 2.5)),
  sex = sample(
    c("Male", "Female", "male", "female", "m", "f", "M", "F"), 50, TRUE
  ),
  age_yrs = pmin(pmax(rnorm(50, 10, 8), 2), 19.5)
)

## WHO method warns that code is experimental
suppressWarnings(weight_status(
  method = "WHO youth",
  bmi = as_bmi(rnorm(50, 23, 2.5)),
  sex = sample(
    c("Male", "Female", "male", "female", "m", "f", "M", "F"), 50, TRUE
  ),
  age_yrs = pmin(pmax(rnorm(50, 10, 8), 0), 19)
))

weight_status(
  as_bmi(c(18, 20, 26, 31, 36, 42)), method = "custom",
  select = "classification", labels = NULL, right = TRUE
)

}
\references{
Portions of these functions were developed with reference to public domain
resources provided by the Centers for Disease Control and Prevention, as well
as the World Health Organization. For more information, see:

\url{https://www.cdc.gov/obesity/basics/adult-defining.html}

\url{https://www.cdc.gov/obesity/childhood/defining.html}

\url{https://www.cdc.gov/healthyweight/assessing/bmi/childrens_bmi/tool_for_schools.html}

\url{https://www.who.int/toolkits/child-growth-standards/standards/body-mass-index-for-age-bmi-for-age}
}
\seealso{
\href{https://academic.oup.com/ajcn/article/90/5/1314/4598130}{Flegal et al. (2009)}
\href{https://www.ahajournals.org/doi/10.1161/CIR.0b013e3182a5cfb3}{Kelly et al. (2013)}
}
