% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_status_CDC_youth.R,
%   R/weight_status_WHO_youth.R, R/weight_status__main.R,
%   R/weight_status_adult_&_custom.R
\name{weight_status_cdc_youth}
\alias{weight_status_cdc_youth}
\alias{weight_status_who_youth}
\alias{weight_status}
\alias{weight_status_adult}
\alias{weight_status_custom}
\title{Classify weight status}
\usage{
weight_status_cdc_youth(bmi)

weight_status_who_youth(bmi)

weight_status(
  bmi,
  wt = NULL,
  ht = NULL,
  wt_units = "kg",
  ht_units = "cm",
  method = c("adult", "CDC youth", "WHO youth", "custom"),
  select = c("bmi", "percentile", "classification", "all"),
  ...
)

weight_status_adult(bmi)

weight_status_custom(
  bmi,
  cutoffs = c(-Inf, 18.5, 25, 30, 35, 40, Inf),
  labels = c("Underweight", "Healthy Weight", "Overweight", "Class 1 Obese",
    "Class 2 Obese", "Class 3 Obese"),
  right = FALSE,
  check_cutoffs = TRUE
)
}
\arguments{
\item{bmi}{body mass index value(s). Must return \code{TRUE} when tested with
\code{\link{is_bmi}}. See details}

\item{wt}{body mass value(s)}

\item{ht}{height value(s)}

\item{wt_units}{units of measure for \code{wt}. Can be quoted (recommended)
or unquoted. Default is kg}

\item{ht_units}{units of measure for \code{ht}. Can be quoted (recommended)
or unquoted. Default is cm}

\item{method}{character scalar. The desired classification method. Options
are \code{"adult"} (the default), \code{"CDC youth"}, \code{"WHO youth"},
and \code{"custom"}}

\item{select}{character. The desired output. Can be \code{"bmi"},
\code{"percentile"}, \code{"classification"}, \code{"all"}, or a
combination thereof. Percentiles will not be returned when \code{method} is
\code{adult} or \code{custom}, and an error will be thrown if no other
selections are made for those methods}

\item{...}{arguments passed to the internal applicator that corresponds with \code{method}}

\item{cutoffs}{The BMI cutoffs to use when determining weight status for
\code{weight_status_custom}. Default is the same cutoffs used for
\code{weight_status_adult}}

\item{labels}{Character labels to assign for intervals based on
\code{cutoffs}. Must be either \code{NULL} or else a character vector of
length one less than \code{length(cutoffs)}. See ?cut for more information}

\item{right}{See ?cut}

\item{check_cutoffs}{logical. Should custom cutoffs be checked with warnings
issued when potential issues are detected?}
}
\value{
A tibble indicating the BMI(s), percentile(s), and/or classification(s)
}
\description{
Classify weight status
}
\details{
Two options are available for passing in values. The first is to
  supply a value for \code{bmi}, which must return \code{TRUE} when tested
  with \code{\link{is_bmi}}. If attempting to pass in BMI as a numeric value,
  it can be wrapped in \code{\link{as_bmi}} to ensure success. The second
  option is to pass weight and height in as separate values, which will be
  fed into \code{\link{get_bmi}} to determine the value for further
  classification. Regardless of how values are passed in, careful attention
  should be given to ensure units are correct. The code and internal checks
  have been set up to help with this as much as possible, but a good amount
  of onus is still on the user and can likely never be completely eliminated.
}
\examples{
weight_status(as_bmi(c(18, 20, 26, 31, 36, 42)))
weight_status(
  as_bmi(c(18, 20, 26, 31, 36, 42)),
  method = "custom", labels = NULL, right = TRUE
)

}
\keyword{internal}
